requirements.txt
asgiref==3.10.0
Django==5.2.7
django-crontab==0.7.1
django-unfold==0.68.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
